requirements.txt
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
autograd==1.8.0
autograd-gamma==0.5.0
biopython==1.85
biotite==1.4.0
biotraj==1.2.2
boto3==1.40.33
botocore==1.40.33
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
cloudpathlib==0.22.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
dna_features_viewer==3.1.5
einops==0.8.1
executing==2.2.1
fair-esm==2.0.0
filelock==3.13.1
fonttools==4.60.0
formulaic==1.2.0
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
interface-meta==1.3.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.2
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lifelines==0.30.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.1
msgpack-numpy==0.4.8
narwhals==2.5.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
py3Dmol==2.5.2
pydssp==0.9.1
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
tdqm==0.0.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchtext==0.18.0
torchvision==0.21.0+cu124
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.3
zstd==1.5.7.2
